+++ components/home-components/banner.tsx
@@ -59,6 +59,12 @@ export default function Banner({ data }: HomePageType) {
 
           button.getToKnowUsButton {
             background: #4e47f9;
+            transition: all 0.3s ease-in-out;
+          }
+          button.getToKnowUsButton:hover {
+            background: white;
+            color: #4e47f9;
+            transition: all 0.3s ease-in-out;
           }
           @media screen and (max-width: 768px) {
             div.image {
