+++ package.json
@@ -9,6 +9,7 @@
     "postinstall": "lefthook install || true"
   },
   "dependencies": {
+    "@types/react-burger-menu": "^2.8.3",
     "formik": "^2.2.9",
     "graphql": "^16.2.0",
     "graphql-request": "^3.7.0",
@@ -16,6 +17,7 @@
     "next-plugin-preval": "^1.2.4",
     "prettier": "^2.5.1",
     "react": "17.0.2",
+    "react-burger-menu": "2.7.1",
     "react-dom": "17.0.2",
     "smoothscroll-polyfill": "^0.4.4",
     "tailwindcss": "^3.0.15",
